requirements.txt
comtypes==1.2.0
ewmh==0.1.6
netifaces==0.11.0
pathlib==1.0.1
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
PyRect==0.2.0
python-xlib==0.33
pywinauto==0.6.6
six==1.16.0
